+++ CHANGELOG.md
@@ -4,7 +4,12 @@
 ## Changed
 - fix opening a shared link when group access to the app is restricted
 - Nextcloud v30 is no longer supported
+- delete extra permissions on file delete
+- replaced general save success message with dedicated messages for each settings section
 - show warning popup when jwt secret is empty and success message otherwise
+
+## Added
+- plugin description and useful links in admin settings
 
 ## 9.11.0
 ## Changed